fix arbitrum deployment scripts
